Privacy page mobile is done
